Provide MetricsSource operator (still under construction)
Add the com.ibm.streamsx.metrics toolkit that contains a single Java
operator, the MetricsSource operator.

The MetricsSource operator captures operator (and - in future - port)
metrics. It detects existing instances, jobs, etc. during the startup
phase and registers itself to receive JMX notifications for new/deleted
instances and jobs, so these will participate - in future - the metrics
capturing.

All jobs that match certain filter criteria are scanned periodically for
changed metric values. For each changed metric value a tuple is emitted
that contains the domain, instance, job, operator, and metric name, as
well as the job id and the metric value.

The toolkit and the Java operator are currently build with Streams
Studio.

This toolkit and its operator are still under construction and will
probably undergo several major changes.