Fixed dynamic detection and custom metrics. Support lastTimeRetrieved.
The operator now detects new instances and jobs as expected. To compare
strings the String.equals method must be used instead of "==".

Custom metrics can be created in an operator's state, onTuple, or
onPunct clause. Therefore, these metrics are not automatically available
as soon as the job is created. Because of missing CUSTOM_METRIC_CREATED
notifications that are available in OperatorContextMXBean only (not
accessible via JMX interface), the operator now keeps track of custom
metrics with a map for captured, and a set for uncaptured metrics. If a
metric is encountered that is neither in the map, nor the set, the
filter matching runs to decide where to put the unhandled metric - into
the map, or the set.

Metrics have the lastTimeRetrieved value, which can now be stored in the
lastTimeRetrieved attribute.