Provide metric type and kind in output tuple
If the `rstring metricType` attribute is part of the output schema, the
metric type, which can be, for example, system or custom, is stored in
this attribute.

If the `rstring metricKind` attribute is part of the output schema, the
metric type, which can be unknown, counter, gauge, or time, is stored in
this attribute.