[ISSUE 17] Make MetricsSource relocatable
Using the @Libraries annotation to specify used class libraries is not
recommended if the class library paths contain an environment variable
like STREAMS_INSTALL, because the @Libraries annotation is evaluated
during compile-time only. This means that the libraries must be stored
in the run-time environment in the same location (same path) as in the
build environment.

Therefore, the used libraries are specified during run-time, during the
operator's initialization. The current value of the environment variable
is used to build the paths to the libraries.